+++ src/test/java/io/beanmapper/testmodel/enums/Day.java
@@ -1,0 +1,24 @@
+package io.beanmapper.testmodel.enums;
+
+public enum Day {
+
+    MONDAY("Monday"),
+    TUESDAY("Tuesday"),
+    WEDNESDAY("Wednesday"),
+    THURSDAY("Thursday"),
+    FRIDAY("Friday"),
+    SATURDAY("Saturday"),
+    SUNDAY("Sunday");
+
+    private final String displayName;
+
+    Day(String displayName) {
+        this.displayName = displayName;
+    }
+
+    @Override
+    public String toString() {
+        return this.displayName;
+    }
+
+}