+++ src/test/java/io/beanmapper/testmodel/enums/WeekEntity.java
@@ -1,0 +1,14 @@
+package io.beanmapper.testmodel.enums;
+
+public class WeekEntity {
+
+    private Day currentDay;
+
+    public Day getCurrentDay() {
+        return currentDay;
+    }
+
+    public void setCurrentDay(Day day) {
+        this.currentDay = day;
+    }
+}